move api helper to core module
